requirement.txt
absl-py==1.4.0
aiofiles==23.1.0
altgraph==0.17.3
anyio==3.6.2
astunparse==1.6.3
audioread==3.0.0
autopep8==2.0.1
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.8
configparser==5.3.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
editdistance==0.6.2
einops==0.2.0
etelemetry==0.3.0
filelock==3.11.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==2.0.7
fonttools==4.39.3
fpdf==1.7.2
frontend==0.0.3
future==0.18.3
gast==0.4.0
gdown==4.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.4
imageio==2.27.0
imgaug==0.4.0
isodate==0.6.1
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.9.2
llvmlite==0.39.1
lmdb==1.4.1
looseversion==1.1.2
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mltu==1.0.7
mpmath==1.3.0
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
onnx==1.12.0
onnxruntime==1.14.1
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.0
patch-ng==1.17.4
path==16.6.0
pathlib==1.0.1
pefile==2023.2.7
Pillow==9.4.0
platformdirs==3.2.0
pooch==1.7.0
prefetch-generator==1.0.1
protobuf==3.20.1
prov==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydot==1.4.2
pyenchant==3.2.2
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypi==2.1
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
pyreadline3==3.4.1
PySide2==5.15.2.1
PySocks==1.7.1
pytesseract==0.3.10
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytils==0.4.1
pytz==2023.3
pyvi==0.1.1
PyWavelets==1.4.1
pywin32-ctypes==0.2.0
pyxnat==1.5
PyYAML==6.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
rdflib==6.3.2
regex==2022.10.31
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
shiboken2==5.15.2.1
simplejson==3.19.1
six==1.16.0
sklearn-crfsuite==0.3.6
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tf2onnx==1.14.0
tflearn==0.5.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
tools==0.1.9
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tqdm==4.64.1
traits==6.3.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
vietocr==0.3.11
Werkzeug==2.2.3
wrapt==1.14.1